+++ packages/start-server-core/src/tanstack-start.d.ts
@@ -1,7 +1,7 @@
 declare module 'tanstack-start-manifest:v' {
   import type { ServerManifest } from '@tanstack/router-core'
 
-  export const tsrStartManifest: () => ServerManifest & { clientEntry: string }
+  export const tsrStartManifest: () => ServerManifest
 }
 
 declare module 'tanstack-start-route-tree:v' {