+++ packages/router-cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tanstack/router-cli
 
+## 1.167.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.167.12
+
+## 1.167.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.167.11
+
 ## 1.167.10
 
 ### Patch Changes