+++ packages/router-plugin/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tanstack/router-plugin
 
+## 1.168.13
+
+### Patch Changes
+
+- Updated dependencies [[`2f53749`](https://github.com/TanStack/router/commit/2f5374945e2138559a51464f45a5152eae67e1dd)]:
+  - @tanstack/router-core@1.171.8
+  - @tanstack/react-router@1.170.10
+  - @tanstack/router-generator@1.167.12
+
+## 1.168.12
+
+### Patch Changes
+
+- Updated dependencies [[`d1997b6`](https://github.com/TanStack/router/commit/d1997b66d7c24c1d64772bb8bab5caf9c6d9cc48)]:
+  - @tanstack/react-router@1.170.9
+  - @tanstack/router-core@1.171.7
+  - @tanstack/router-generator@1.167.11
+
 ## 1.168.11
 
 ### Patch Changes