+++ packages/router-vite-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tanstack/router-vite-plugin
 
+## 1.167.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-plugin@1.168.13
+
+## 1.167.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-plugin@1.168.12
+
 ## 1.167.11
 
 ### Patch Changes