+++ packages/start-client-core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @tanstack/start-client-core
 
+## 1.170.6
+
+### Patch Changes
+
+- Updated dependencies [[`2f53749`](https://github.com/TanStack/router/commit/2f5374945e2138559a51464f45a5152eae67e1dd)]:
+  - @tanstack/router-core@1.171.8
+  - @tanstack/start-storage-context@1.167.10
+
+## 1.170.5
+
+### Patch Changes
+
+- Updated dependencies [[`d1997b6`](https://github.com/TanStack/router/commit/d1997b66d7c24c1d64772bb8bab5caf9c6d9cc48)]:
+  - @tanstack/router-core@1.171.7
+  - @tanstack/start-storage-context@1.167.9
+
 ## 1.170.4
 
 ### Patch Changes