+++ packages/vue-start-client/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tanstack/vue-start-client
 
+## 1.167.11
+
+### Patch Changes
+
+- Updated dependencies [[`2f53749`](https://github.com/TanStack/router/commit/2f5374945e2138559a51464f45a5152eae67e1dd)]:
+  - @tanstack/router-core@1.171.8
+  - @tanstack/start-client-core@1.170.6
+  - @tanstack/vue-router@1.170.10
+
+## 1.167.10
+
+### Patch Changes
+
+- Updated dependencies [[`d1997b6`](https://github.com/TanStack/router/commit/d1997b66d7c24c1d64772bb8bab5caf9c6d9cc48)]:
+  - @tanstack/router-core@1.171.7
+  - @tanstack/vue-router@1.170.9
+  - @tanstack/start-client-core@1.170.5
+
 ## 1.167.9
 
 ### Patch Changes